Add support for "slow stations".
When defining a slow station, trains that don't call at this station will be marked as "fast". Currently there's no UI to show these fast trains.

NB: You need to use the custom webserver that adds the required "callingAt" field, which is too complex for the poor Pico to do on-device.